Revert "Feat(firewall & reverseProxy): Separate firewall component into waf a…"
This reverts commit d843c2acb81f78712a8b078ae05895563dd10499.
