+++ firewall/go.mod
@@ -1,0 +1,3 @@
+module cnss/firewall
+
+go 1.24.1
