--- firewall/reverceProxy/go.mod
@@ -1,3 +1,0 @@
-module cnss/firewall/reverceProxy
-
-go 1.24.1
