--- firewall/waf/go.mod
@@ -1,3 +1,0 @@
-module cnss/firewall/waf
-
-go 1.24.1
